backend/flask/requirements.txt
Flask
Flask-Cors
joblib
pandas
numpy
scikit-learn
mysql-connector-python
shap
gunicorn